requirements.txt
discord.py==2.3.2
yt-dlp==2024.3.10
PyNaCl==1.5.0
ffmpeg-python==0.2.0